refactor(migration-analyzer): DumpFileAnalyzer를 별도 파일로 분리
대형 파일 정리 8번째(Phase H). migration_analyzer.py(1605줄)를 분석한
결과 MigrationAnalyzer(1108줄, 스키마 분석)와 DumpFileAnalyzer(313줄,
덤프파일 분석)는 서로 다른 모델을 쓰는 완전히 독립된 두 서브시스템이었음
(둘 다 이 파일 안에서 서로를 생성하지 않음).

- migration_dump_analyzer.py (신규): DumpAnalysisResult, DumpFileAnalyzer
- migration_analyzer.py: 1605 -> 1285줄 (재수입, IssueType/CompatibilityIssue
  등 외부에서 이 모듈을 통해 재수입해가는 기존 관례 유지)

구현 중 발견한 실제 결합: DumpFileAnalyzer._analyze_sql_file()이 예약어
충돌 검사에 MigrationAnalyzer.NEW_RESERVED_KEYWORDS 클래스 상수를 직접
참조하고 있어서, 단순 이동 시 순환 import(migration_analyzer <->
migration_dump_analyzer)가 발생함. 이 코드베이스에 이미 있는 패턴(지연
import)을 따라 해당 메서드 내부에 `from src.core.migration_analyzer
import MigrationAnalyzer`를 지연 import로 추가해 해결.

검증: 관련 pytest 58 passed (수정 전에는 순환참조로 NameError가 발생해
1개 테스트 조용히 실패 중이었음 — except Exception에 삼켜짐). 전체
pytest 1809 passed / 1 known-flaky 제외. py_compile 통과. 앱 스모크 성공.
